🔧 build(APC): Domain
Organized how folders are structured inside the APC repo folder
