+++ APC/.vscode/settings.json
@@ -1,5 +1,5 @@
 {
-	"python.testing.pytestArgs": ["tests", "-v"],
+	"python.testing.pytestArgs": ["tests/", "-v"],
 	"python.testing.unittestEnabled": false,
 	"python.testing.pytestEnabled": true,
 	"cSpell.words": ["celsius"]
