+++ APC/codes/challenges/painting.py
