+++ APC/codes/linf/triangle.py
